to the rescue with a real linux fix
Apparently the zuul is correctly (mostly) linked to the frontend
container by docker-compose. The really weird thing is though that the
port mapping isn't done or at least not visible in the frontend
container but from the host it works perfectly fine. Therefore accessing
the url "http://zuul:8081" with the feign clients works like a charm.

tl;dr
Why this works? I have a couple of speculations, none of them sound very
likely.
1. Magic
2. Weirdly while building the frontend container exposes the port 8080.
(probably configured invisibly in a parent Dockerfile; or not configured
at all). After that maybe the container cannot map the zuul container to
8080 as it is exposed by the frontend.
3. This is the planned functionality of the link property of
docker-compose, as it only maps the port to the host system and not
between the containers. Previously this maybe worked as the portmapping
was on 8080:8080 or localhost was correctly resolved to the hosts
localhost. Weirdly this is for me the most likely